Worked on patien view (AppintmentList,TreatmentList and Workedone List)

+++ DCRM.Repository/IRepository/IRepository.cs
@@ -10,7 +10,7 @@ namespace DCRM.Repository.IRepository
     public interface IRepository<T> where T : BaseEntity
     {
         IEnumerable<T> GetAll();
-        T Get(int Id);
+        T Get(long Id);
         void Insert(T entity);
         void Update(T entity);
         void Delete(T entity);
